Remove "reset" method of ScheduledTaskRoutine to fix bugs (#369)
